+++ pact/base.py
@@ -4,7 +4,6 @@ import logbook
 import sys
 import waiting
 from ._compat import reraise
-from logbook.utils import deprecated
 from waiting.exceptions import TimeoutExpired
 
 _logger = logbook.Logger(__name__)
@@ -69,13 +68,6 @@ class PactBase(object):
                 _logger.debug("Exception thrown from '{}' callback {!r} of {!r}", name, callback, self, exc_info=True)
         return exc_info
 
-    @deprecated('Use poll() and/or is_finished() instead')
-    def finished(self):
-        """Deprecated. Use poll() or is_finished() instead
-        """
-        self.poll()
-        return self.is_finished()
-
     def then(self, callback, *args, **kwargs):
         """Calls ``callback`` when this pact is finished
         """
@@ -114,24 +106,27 @@ class PactBase(object):
     def wait(self, **kwargs):
         """Waits for this pact to finish
         """
-        _logger.debug("Waiting for {0!r}", self)
+        _logger.debug("Waiting for {!r}...", self)
+        start_time = flux.current_timeline.time()
+        if 'timeout_seconds' not in kwargs and self._end_time is not None:
+            kwargs['timeout_seconds'] = max(0, self._end_time - start_time)
         try:
-            if 'timeout_seconds' not in kwargs and self._end_time is not None:
-                kwargs['timeout_seconds'] = max(0, self._end_time - flux.current_timeline.time())
             waiting.wait(self.poll, waiting_for=self, **kwargs)
-            _logger.debug("Finish waiting for {0!r}", self)
         except TimeoutExpired:
             exc_info = sys.exc_info()
             for timeout_callback in self._timeout_callbacks:
                 timeout_callback()
-            exc = self.get_timeout_exception(exc_info)
+            exc = self.get_timeout_exception(exc_info)  # pylint: disable=assignment-from-none
             if exc is None:
                 reraise(*exc_info)
             else:
                 raise exc # pylint: disable=raising-bad-type
         except Exception:
-            _logger.debug("Exception was raised while waiting for {0!r}", self, exc_info=True)
+            _logger.debug("Exception was raised while waiting for {!r}", self, exc_info=True)
             raise
+        else:
+            finish_time = flux.current_timeline.time()
+            _logger.debug("Finished waiting for {!r} (took: {:.04f} sec)", self, finish_time - start_time)
 
     def __repr__(self):
         raise NotImplementedError() # pragma: no cover